refactor: extract hero effects from AgentWorkspace
