+++ src/components/agent/use-agent-hero-rotation-types.ts
@@ -1,0 +1,9 @@
+import type * as React from "react";
+
+export type useAgentHeroRotationDependencies = {
+  draftRef: React.MutableRefObject<string>;
+  heroChipsHoverRef: React.MutableRefObject<boolean>;
+  heroMode: boolean;
+  setHeroChipPhase: React.Dispatch<React.SetStateAction<"in" | "out">>;
+  setHeroDeckStart: React.Dispatch<React.SetStateAction<number>>;
+};
